Use aruba's environment block to set the test env's PWD
This is needed to allow testing to pass due to changes introduced
in CHEF-3663.

Since the master task process executes in the project's root
directory (PWD) and Chef now looks at ENV['PWD'] for configs,
changing this per test is necessary to ensure the test can load
the supplied config that is generated in aruba's `current_dir`.